fix(android): 支持ARM64运行

+++ cocos2dx/textures/CCTexture2D.cpp
@@ -807,7 +807,12 @@ void CCTexture2D::setAliasTexParameters()
 
     glTexParameteri( GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST );
 #if CC_ENABLE_CACHE_TEXTURE_DATA
-    ccTexParams texParams = {m_bHasMipmaps?GL_NEAREST_MIPMAP_NEAREST:GL_NEAREST,GL_NEAREST,GL_NONE,GL_NONE};
+    ccTexParams texParams = {
+        static_cast<GLuint>(m_bHasMipmaps ? GL_NEAREST_MIPMAP_NEAREST : GL_NEAREST),
+        static_cast<GLuint>(GL_NEAREST),
+        static_cast<GLuint>(GL_NONE),
+        static_cast<GLuint>(GL_NONE)
+    };
     VolatileTexture::setTexParameters(this, &texParams);
 #endif
 }
@@ -827,7 +832,12 @@ void CCTexture2D::setAntiAliasTexParameters()
 
     glTexParameteri( GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR );
 #if CC_ENABLE_CACHE_TEXTURE_DATA
-    ccTexParams texParams = {m_bHasMipmaps?GL_LINEAR_MIPMAP_NEAREST:GL_LINEAR,GL_LINEAR,GL_NONE,GL_NONE};
+    ccTexParams texParams = {
+        static_cast<GLuint>(m_bHasMipmaps ? GL_LINEAR_MIPMAP_NEAREST : GL_LINEAR),
+        static_cast<GLuint>(GL_LINEAR),
+        static_cast<GLuint>(GL_NONE),
+        static_cast<GLuint>(GL_NONE)
+    };
     VolatileTexture::setTexParameters(this, &texParams);
 #endif
 }
